[CHORE] Remove StarStorm README section
