+++ frontend/package.json
@@ -18,6 +18,7 @@
     "@sveltejs/vite-plugin-svelte": "^6.0.0",
     "eslint": "^9.34.0",
     "globals": "^16.3.0",
+    "jsdom": "^26.1.0",
     "svelte": "^5.0.0",
     "vite": "^7.0.4",
     "vite-plugin-static-copy": "^3.1.2"
