+++ frontend/src/lib/components/MenuPanel.svelte
@@ -1,13 +1,17 @@
 <script>
+  import StarStorm from './StarStorm.svelte';
   export let padding = '0.5rem';
+  export let reducedMotion = false;
+  export let starColor = '';
 </script>
 
 <style>
   .panel {
-  width: 100%;
-  height: 100%;
-  max-width: 100%;
-  max-height: 100%;
+    position: relative;
+    width: 100%;
+    height: 100%;
+    max-width: 100%;
+    max-height: 100%;
     display: flex;
     flex-direction: column;
     box-sizing: border-box;
@@ -42,6 +46,7 @@
   }
 </style>
 
-<div class="panel" style={`--padding: ${padding}`}> 
+<div class="panel" style={`--padding: ${padding}`}>
+  <StarStorm color={starColor} {reducedMotion} />
   <slot />
 </div>
