+++ frontend/src/lib/components/PartyPicker.svelte
@@ -1,5 +1,5 @@
 <script>
-  import { onMount, onDestroy } from 'svelte';
+  import { onMount } from 'svelte';
   import { createEventDispatcher } from 'svelte';
   import { getPlayers } from '../systems/api.js';
   import { getCharacterImage, getHourlyBackground, getRandomFallback, getElementColor } from '../systems/assetLoader.js';
@@ -29,113 +29,12 @@
     if (!cur?.is_player) previewElementOverride = '';
   }
 
-  // Starfield background tied to current preview element color
-  function rand(min, max) { return Math.random() * (max - min) + min; }
-  const starElements = ['fire','ice','lightning','light','dark','wind'];
-  function randomElementColor() {
-    const el = starElements[Math.floor(Math.random() * starElements.length)];
-    try { return getElementColor(el); } catch { return '#88a'; }
-  }
-  function spawnStar() {
-    const c = randomElementColor();
-    return {
-      left: Math.random() * 100,
-      size: rand(3, 6),
-      duration: rand(6, 14),
-      delay: rand(0, 6),
-      drift: rand(-20, 20),
-      color: c
-    };
-  }
-  function makeStars(count) {
-    return Array.from({ length: count }, () => spawnStar());
-  }
-  $: density = reducedMotion ? 60 : 240;
-  let stars = [];
   $: currentElementName = (() => {
     const cur = roster.find(r => r.id === previewId);
     const el = previewElementOverride || (cur && cur.element) || '';
     return el ? String(el) : '';
   })();
   $: starColor = currentElementName ? (() => { try { return getElementColor(currentElementName); } catch { return ''; } })() : '';
-
-  // Global fade pulse helper
-  let fading = false;
-  function fadePulse(dur = 260, low = 0.25) {
-    // use CSS transition on .stars to dip opacity, then restore
-    fading = true;
-    clearTimeout(fadeTimer);
-    fadeTimer = setTimeout(() => { fading = false; }, dur);
-  }
-  let fadeTimer;
-
-  // Initialize stars regardless of element; each star uses a random element color
-  $: if (stars.length === 0) {
-    stars = makeStars(density);
-    fadePulse(260);
-  }
-
-  // Adjust star count when density changes
-  $: if (stars.length && stars.length !== density) {
-    if (density > stars.length) {
-      stars = stars.concat(makeStars(density - stars.length));
-    } else if (density < stars.length) {
-      stars = stars.slice(0, density);
-    }
-  }
-
-  // Gradual color transition: on change, slowly replace existing stars
-  let lastStarColor = '';
-  let replaceTimer = null;
-  let forceTimer = null;
-  function startColorTransition() {
-    if (!starColor) return;
-    if (replaceTimer) clearInterval(replaceTimer);
-    if (forceTimer) clearTimeout(forceTimer);
-    fadePulse(260);
-    // Build a deterministic pool of indices that still have the old color
-    let pool = stars
-      .map((s, i) => ({ s, i }))
-      .filter(({ s }) => s.color !== starColor)
-      .map(({ i }) => i);
-    if (pool.length === 0) return;
-    // Compute batch size to finish around target duration
-    const intervalMs = 120;
-    const targetMs = 1800;
-    const ticks = Math.max(1, Math.round(targetMs / intervalMs));
-    const batch = Math.max(1, Math.ceil(pool.length / ticks));
-    replaceTimer = setInterval(() => {
-      const n = Math.min(batch, pool.length);
-      for (let k = 0; k < n; k++) {
-        const idx = pool.pop();
-        if (idx == null) break;
-        // Replace with a fresh star using a random element color
-        stars[idx] = spawnStar();
-      }
-      stars = stars.slice();
-      if (pool.length === 0) {
-        clearInterval(replaceTimer);
-        replaceTimer = null;
-      }
-    }, intervalMs);
-    // Safety: force any remaining to new color shortly after target
-    forceTimer = setTimeout(() => {
-      if (pool.length > 0) {
-        stars = stars.map(s => (s.color === starColor ? s : spawnStar()));
-      }
-      if (replaceTimer) { clearInterval(replaceTimer); replaceTimer = null; }
-    }, targetMs + 400);
-  }
-  $: if (starColor && lastStarColor && starColor !== lastStarColor) {
-    startColorTransition();
-  }
-  $: lastStarColor = starColor;
-
-  onDestroy(() => {
-    if (replaceTimer) clearInterval(replaceTimer);
-    if (fadeTimer) clearTimeout(fadeTimer);
-    if (forceTimer) clearTimeout(forceTimer);
-  });
 
   onMount(async () => {
     background = getHourlyBackground();
@@ -192,16 +91,8 @@
 {#if compact}
   <PartyRoster {roster} {selected} bind:previewId {compact} {reducedMotion} on:toggle={(e) => toggleMember(e.detail)} />
 {:else}
-  <MenuPanel style={`background-image: url(${background}); background-size: cover;`}>
+  <MenuPanel style={`background-image: url(${background}); background-size: cover;`} {starColor} {reducedMotion}>
     <div class="full" data-testid="party-picker">
-      <!-- Starfield background layer (behind content) -->
-      <div class="stars" aria-hidden="true" style={`opacity:${fading ? 0.25 : 0.55}`}> 
-        {#each stars as s}
-          <span class="star" style={`--x:${s.left}%; --s:${s.size}px; --d:${s.duration}s; --delay:${s.delay}s; --dx:${s.drift}px; --c:${s.color};`}>
-            <span class="core"></span>
-          </span>
-        {/each}
-      </div>
       <PartyRoster {roster} {selected} bind:previewId {reducedMotion} on:toggle={(e) => toggleMember(e.detail)} />
       <PlayerPreview {roster} {previewId} overrideElement={previewElementOverride} />
       <div class="right-col">
@@ -286,61 +177,4 @@
   .party-actions-inline .wide:hover { background: rgba(255,255,255,0.1); border-color: rgba(255,255,255,0.5); }
 
   /* Falling starfield */
-  .stars { position:absolute; inset:0; overflow:hidden; pointer-events:none; z-index:-1; opacity: 0.55; transition: opacity 220ms ease; }
-  /* Parent handles vertical fall + opacity */
-  .star {
-    position: absolute;
-    top: -10%;
-    left: var(--x);
-    width: 0;  /* visuals live in .core */
-    height: 0; /* prevents layout shifts */
-    animation-name: af-fallTop;
-    animation-timing-function: linear;
-    animation-iteration-count: infinite;
-    animation-duration: var(--d);
-    animation-delay: var(--delay);
-  }
-  /* Child handles horizontal drift and the visual shape */
-  .star .core {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: var(--s);
-    height: var(--s);
-    background: radial-gradient(circle, var(--c) 0%, transparent 70%);
-    border-radius: 50%;
-    filter: drop-shadow(0 0 8px color-mix(in srgb, var(--c) 70%, transparent));
-    animation-name: af-drift;
-    animation-timing-function: linear;
-    animation-iteration-count: infinite;
-    animation-duration: var(--d);
-    animation-delay: var(--delay);
-  }
-  .star .core::after {
-    content: '';
-    position: absolute;
-    left: 50%;
-    transform: translateX(-50%);
-    top: calc(var(--s) * -8);
-    width: calc(max(2px, var(--s) * 0.4));
-    height: calc(var(--s) * 10);
-    background: linear-gradient(180deg, color-mix(in srgb, var(--c) 75%, transparent) 0%, transparent 70%);
-    filter: blur(1px);
-  }
-  @keyframes af-fallTop {
-    0% { top: -10%; opacity: 0.0; }
-    10% { opacity: 1.0; }
-    90% { opacity: 1.0; }
-    100% { top: 110%; opacity: 0.0; }
-  }
-  @keyframes af-drift {
-    0% { transform: translateX(0); }
-    100% { transform: translateX(var(--dx)); }
-  }
-
-  /* Respect reduced motion */
-  :global(html.reduced-motion) .stars .star,
-  :global(body.reduced-motion) .stars .star { animation: none; opacity: 0.35; top: 15%; }
-  :global(html.reduced-motion) .stars .star .core,
-  :global(body.reduced-motion) .stars .star .core { animation: none; }
   </style>
